Add files via upload
This model uses a convolutional neural network architecture with multiple convolutional and pooling layers, followed by fully connected layers. It takes in input images of faces and outputs the probabilities of "happy" and "sad" emotions.
The description of the deep CNN model for detecting happy and sad emotions was inspired by Nicholas Renotte's teachings on convolutional neural networks.